+++ executor/Cargo.toml
@@ -2,12 +2,17 @@
 name = "cennznet-executor"
 version = "1.0.0"
 authors = ["Centrality Developers <support@centrality.ai>"]
+description = "CENNZnet node implementation in Rust."
 edition = "2018"
+license = "GPL-3.0"
+homepage = "https://cennznetdocs.com"
+repository = "https://github.com/cennznet/cennznet"
 
 [dependencies]
 # third-party dependencies
 codec = { version = "1.2.0", package = "parity-scale-codec" }
-trie-root = "0.15.2"
+trie-root = "0.16.0"
+frame-benchmarking = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
 
 # internal dependencies
 cennznet-primitives = { path = "../primitives" }
@@ -20,16 +25,33 @@ sp-core = { git = "https://github.com/plugblockchain/plug-blockchain", branch = 
 sp-trie = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
 
 [dev-dependencies]
-wabt = "0.9.2"
-cennznet-testing = { path = "../testing" }
-crml-transaction-payment = { path="../crml/transaction-payment", default-features = false }
-sp-runtime = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+criterion = "0.3.0"
 frame-support = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
-pallet-generic-asset = { git = "https://github.com/plugblockchain/plug-blockchain", default-features = false, branch = "develop" }
 frame-system = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
-pallet-timestamp = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+cennznet-testing = { path = "../testing" }
 pallet-contracts = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+pallet-grandpa = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+pallet-im-online = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+pallet-session = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+pallet-timestamp = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+pallet-generic-asset = { git = "https://github.com/plugblockchain/plug-blockchain", default-features = false, branch = "develop" }
+crml-transaction-payment = { path="../crml/transaction-payment", default-features = false }
+pallet-treasury = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+sp-application-crypto = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+sp-runtime = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+substrate-test-client = { git = "https://github.com/plugblockchain/plug-blockchain", branch = "develop" }
+wabt = "0.9.2"
+
 
 [features]
-benchmarks = []
+wasmtime = [
+	"sc-executor/wasmtime",
+]
+wasmi-errno = [
+	"sc-executor/wasmi-errno",
+]
 stress-test = []
+
+[[bench]]
+name = "bench"
+harness = false
